Now switch back to Event based client updates if clientUpdateRate is 0 or omitted. Fixed issue with updates not being sent if arrays or hashmaps were updated.
