+++ conf/application-prod.conf
@@ -6,7 +6,8 @@
 breadboard.amtFrameHeight=600
 
 # How frequently, in milliseconds, to send updates to the client
-breadboard.clientUpdateRate=1000
+# If 0, or omitted, default to v2.2 event based updates
+breadboard.clientUpdateRate=0
 
 breadboard.client.messageBuffer=25000000
